requirements.txt
discord.py>=2.4.0
httpx>=0.27.0
asyncpg>=0.29.0
python-dotenv>=1.0.1
